+++ gatsby-node.js
@@ -1,6 +1,7 @@
-const path = require("path")
-var fs = require("fs")
-var dir = "./.cache/caches/gatsby-source-prismic-graphql"
+const fs = require("fs")
+const dir = "./.cache/caches/gatsby-source-prismic-graphql"
+const { createInsights } = require("./templates/createInsights")
+const { createPosts } = require("./templates/createPosts")
 
 /* Fix for gatsby-source-prismic-graphql bug related to imageSharp
 https://github.com/birkir/gatsby-source-prismic-graphql/issues/162 */
@@ -8,64 +9,6 @@ exports.onPreBootstrap = () => {
   if (!fs.existsSync(dir)) {
     fs.mkdirSync(dir)
   }
-}
-
-const { postsPerPage } = require("./config/website")
-
-const createInsights = ({ createPage }) => (result) => {
-  const insightTemplate = path.resolve("src/pages-generic/insight/insight.jsx")
-  const insights = result.data.prismic.allInsights.edges
-
-  insights.forEach((insight) => {
-    const uid = insight.node._meta.uid
-
-    createPage({
-      path: `/insights/${uid}`,
-      component: insightTemplate,
-      context: {
-        uid,
-      },
-    })
-  })
-}
-
-const createPosts = ({ createPage }) => (result) => {
-  const postTemplate = path.resolve("src/pages-generic/post.jsx")
-  const postsTemplate = path.resolve("src/pages-generic/posts.jsx")
-
-  const posts = result.data.prismic.allPosts.edges
-  const totalCount = result.data.prismic.allInsights.totalCount
-
-  posts.forEach((post) => {
-    const uid = post.node._meta.uid
-
-    createPage({
-      path: `/insights/${uid}`,
-      component: postTemplate,
-      context: {
-        uid,
-      },
-    })
-  })
-
-  const numPages = Math.ceil(totalCount / postsPerPage)
-
-  Array.from({ length: numPages }).forEach((_, i) => {
-    const limit = postsPerPage
-    const skip = i * postsPerPage
-    const currentPage = i + 1
-
-    createPage({
-      path: i === 0 ? "/insights" : `/insights/${i + 1}`,
-      component: postsTemplate,
-      context: {
-        limit,
-        skip: limit - skip,
-        numPages,
-        currentPage,
-      },
-    })
-  })
 }
 
 const resultCheck = ({ reporter }) => (result) => {
