+++ src/styles/abstracts/_variables.scss
@@ -1,6 +1,5 @@
 // Google Fonts
-@import url(https://fonts.googleapis.com/css?family=Open+Sans:200,300,400,400i,500,600,700);
-@import url(https://fonts.googleapis.com/css2?family=Poppins&display=swap);
+@import url(https://fonts.googleapis.com/css2?family=Poppins&display=swap:100,200,300,400,500,600,700,800);
 
 // Font variables
 $poppins: 'Poppins', sans-serif;
