+++ src/styles/theme/fonts.jsx
@@ -1,17 +1,17 @@
 export const fonts = {
   fontFamily: "'Poppins', sans-serif",
   fontSize: {
-    h1: {
+    heading: {
       lg: "8rem",
       md: "6rem",
       sm: "4rem",
       xs: "2rem",
     },
-    meta: {
-      lg: "1.5rem",
-      md: "1.25rem",
-      sm: "1rem",
-      xs: "0.75rem",
+    paragraph: {
+      lg: "1.8rem",
+      md: "1.6rem",
+      sm: "1.5rem",
+      xs: "1.4rem",
     },
   },
   fontWeight: {
